requirements.txt
pandas==2.0.3
numpy==1.24.3
regex==2023.6.3
pyyaml==6.0.1
loguru==0.7.0
matplotlib==3.7.2
pytest==7.4.0
